+++ apps/zui/src/zui-kit/index.ts
@@ -1,3 +1,3 @@
 export * from "./core"
-export * from "./react/list-view"
-export * from "./react/table-view"
+export * from "./list/list-view"
+export * from "./table/table-view"
