+++ apps/zui/src/zui-kit/list/list-view.tsx
@@ -12,9 +12,9 @@ import {ListViewApi} from "../core"
 import {createListView} from "../core/list-view/create-list-view"
 import {defaultListViewState} from "../core/list-view/state"
 import {ListViewArgs} from "../core/list-view/types"
-import {ReactAdapterProps} from "./types"
+import {ReactAdapterProps} from "../types/types"
 import {useStateControllers} from "../utils/use-state-controllers"
-import {useInitialScrollPosition, useOnScroll} from "./utils"
+import {useInitialScrollPosition, useOnScroll} from "../utils/utils"
 import classNames from "classnames"
 import {useParentSize} from "../utils/use-parent-size"
