+++ apps/zui/src/zui-kit/table/table-view.tsx
@@ -3,7 +3,7 @@ import {Provider} from "./context"
 import {Grid} from "./grid"
 import classNames from "classnames"
 import {TableViewArgs} from "../core/table-view/types"
-import {ReactAdapterProps} from "../react/types"
+import {ReactAdapterProps} from "../types/types"
 import {defaultTableViewState} from "../core/table-view/state"
 import {TableViewApi} from "./table-view-api"
 import {useStateControllers} from "../utils/use-state-controllers"
